requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
supabase>=2.0.0
pdfplumber>=0.10.0
pdf2image>=1.16.0
python-dotenv>=1.0.0
openai>=1.0.0
Pillow>=10.0.0
httpx>=0.26.0
